+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+Version 208:
+
+* Add get_lowest_layer free function
+* Add lowest_layer_type metafunction
+
+--------------------------------------------------------------------------------
+
 Version 207
 
 * Send from the strand